requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.0.3
SQLAlchemy==2.0.40
cryptography==39.0.2
PyJWT==2.10.1